requirements.txt
Flask==1.1.1
python-dotenv==0.10.3
flask-mongoengine==0.9.5
Flask-RESTful==0.3.7
passlib==1.7.1
Flask-JWT-Extended==3.23.0
